requirements.txt
scrapy>=2.5.0
selenium>=4.1.0
scrapy-selenium>=0.0.7
mysql-connector-python>=8.0.26
redis>=4.0.0
elasticsearch>=8.0.0
elasticsearch-dsl>=8.0.0
cryptography>=3.4.7
pika>=1.3.0
